Better ancestry constraints; README.
(1) I constrained the rates of the branches leading up to languages
constrained to be ancestors to be 1.  I first tried setting these rates
to 0, but the likelihood function gave NaN ("not a number", an error
result) probably due to BEAST not properly handling this edge case.
However, even at a rate of 1, the impact of these very short branches
is negligible, especially given the following change.

(2) These branch rates are no longer drawn from the log-normal
distribution of per-branch rates.  This should result in a much smaller
variance in per-branch rates.

The implementation details are as follows.  In the XML file, for each
taxon specifying an ancient language constrained to be an ancestor,
add an "adventitious" attribute:

        <taxon id="Latin" adventitious="true">
            <date value="2100.0" direction="backwards" units="years"/>
        </taxon>

Also do everything that was described in the last checkin.  I hacked
BEAST so that the getter for relative branch rates will always return
1.0 when it is connected to an adventitious leaf.  Underlyingly there
is a different relative branch rate, sampled from the log-normal, that
is never used.  Since it is never used, it has no effect on the model,
except perhaps to make it somewhat harder for the model to mix.  But rate
variation is so low that I doubt there's even much of an affect there.