Fix covarion model
In the code to construct the unnormalized rate matrix, the hfrequency
parameter is completely ignored.  However, elsewhere, hfrequency is used
to calculate the stationary frequency of all four states.  This means
that the matrix is normalized incorrectly.

Now, hfrequency is no longer ignored in the code to construct the
unormalized rate matrix.