Add startIter key to operators
You can now tell BEAST to not use an operator until the chain length
reaches N.  This can be useful for avoiding local maxima.

If the scheduler picks an operator to run, it is summarily rejected if
the chain length is smaller than the startIter value for the operator.
This value is specified in the XML using the startIter key.

I have no idea how this interacts with optimizing operators, etc.